+++ NEWS.md
@@ -1,5 +1,6 @@
 # aion 1.3.0.9000
 ## New classes and methods
+* Add `as_graph()` to compute an interval graph.
 * Add `get_calendar()` and `set_calendar()` to get and set the default calendar.
 
 # aion 1.3.0