+++ NEWS.md
@@ -1,4 +1,4 @@
-# aion 1.3.0.9000
+# aion 1.4.0
 ## New classes and methods
 * Add `as_graph()` to compute an interval graph.
 * Add `get_calendar()` and `set_calendar()` to get and set the default calendar.